refactor: card submit 로직 hook으로 분리
